+++ CHANGELOG.md
@@ -6,7 +6,7 @@ http://phpsysinfo.sourceforge.net/
 phpSysInfo 3.1.17
 ----------------
 
- - [UPD] BAT plugin - capacity unit, cycle count
+ - [UPD] BAT plugin - capacity unit, cycle count, FreeBSD support
  - [UPD] Brazilian Portuguese Translation pt-br.xml
 
 phpSysInfo 3.1.16